+++ sdk/src/main/java/com/mapbox/maps/viewannotation/ViewAnnotationManager.kt
@@ -156,6 +156,17 @@ interface ViewAnnotationManager {
   fun getViewAnnotationUpdateMode(): ViewAnnotationUpdateMode
 
   /**
+   * Return Map of added [View] with their [ViewAnnotationOptions]. This method should be called every time
+   * to get all added [View]s and associated [ViewAnnotationOptions].
+   *
+   * Note: Modifying [ViewAnnotationOptions] will not update actual annotation options. Use
+   * [ViewAnnotationManager.updateViewAnnotation] instead. Modifying [View] will update the actual view property.
+   *
+   * @return Map of [View] and associated [ViewAnnotationOptions].
+   */
+  val annotations: Map<View, ViewAnnotationOptions>
+
+  /**
    * Static methods and variables.
    */
   companion object {